Clean up interfaces
Introduces StreamDialer and PacketDialer.
