+++ net/stream.go
@@ -1,3 +1,17 @@
+// Copyright 2019 Jigsaw Operations LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package net
 
 import (
@@ -15,6 +29,13 @@ type DuplexConn interface {
 	// Closes the Write end of the connection. An EOF or FIN signal may be
 	// sent to the connection target.
 	CloseWrite() error
+}
+
+// StreamDialer provides a way to establish stream connections (like TCP).
+type StreamDialer interface {
+	// Dial connects to `raddr`.
+	// `raddr` has the form `host:port`, where `host` can be a domain name or IP address.
+	Dial(raddr string) (DuplexConn, error)
 }
 
 type duplexConnAdaptor struct {
@@ -85,15 +106,4 @@ func Relay(leftConn, rightConn DuplexConn) (int64, int64, error) {
 		err = rs.Err
 	}
 	return n, rs.N, err
-}
-
-type ConnectionError struct {
-	// TODO: create status enums and move to metrics.go
-	Status  string
-	Message string
-	Cause   error
-}
-
-func NewConnectionError(status, message string, cause error) *ConnectionError {
-	return &ConnectionError{Status: status, Message: message, Cause: cause}
 }
